Adjusted OpenSUSE cleanup script

+++ scripts/opensuse/cleanup.sh
@@ -1,3 +1,12 @@
 #!/bin/bash -eux
 # These were only needed for building VMware/Virtualbox extensions:
-zypper -n rm -u binutils gcc make perl ruby kernel-default-devel kernel-devel
+
+version=`grep VERSION= /etc/os-release | cut -f2 -d\" | cut -f1 -d\ `
+
+if [[ $version =~ "13" ]]; then
+  zypper -n rm -u binutils gcc make perl ruby kernel-default-devel kernel-devel
+fi
+
+if [[ $version =~ "4" ]]; then
+  zypper -n rm -u gcc make kernel-default-devel kernel-devel
+fi
